+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - UNRELEASED
+## [2.0.0] - UNRELEASED
 ### Added
 - export feature
 - Endpoint to enable importing elements to a staging area
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Endpoint to enable converting the stagedElements to drafts
 - Endpoint to delete an import by ID
 - Endpoint to return all available paths for specified element
+- Endpoint to update the members of dataElementGroup/record
 ### Changed
+- ElementController does not process namespaces anymore 
 ### Deprecated
 ### Removed
 ### Fixed